api/requirements.txt
# Core API dependencies
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-multipart==0.0.6

# SSE support
sse-starlette==1.8.2

# Note: Parent requirements.txt already includes:
# - daytona
# - python-dotenv==1.0.0
# - rich==13.7.0
# - prompt-toolkit>=3.0.0
# - aiohttp>=3.9.0
# - pytest==7.4.3
# - pytest-asyncio==0.21.1

# Testing (API specific)
httpx==0.25.2

# Observability (optional)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0